# [seqbench] `the-good-parts` sequence