--- a/libsolidity/interface/OptimiserSettings.h
+++ b/libsolidity/interface/OptimiserSettings.h
@@ -43,20 +43,21 @@ struct OptimiserSettings
 {
 	static char constexpr DefaultYulOptimiserSteps[] =
 		"dhfoDgvulfnTUtnIf"            // None of these can make stack problems worse
-		"["
-			"xa[r]EscLM"               // Turn into SSA and simplify
-			"cCTUtTOntnfDIul"          // Perform structural simplification
-			"Lcul"                     // Simplify again
-			"Vcul [j]"                 // Reverse SSA
 
-			// should have good "compilability" property here.
+		"xa[r]EscLM"                   // Turn into SSA and simplify
+		"Vcul [j]"                     // Reverse SSA
 
-			"Tpeul"                    // Run functional expression inliner
-			"xa[rul]"                  // Prune a bit more in SSA
-			"xa[r]cL"                  // Turn into SSA again and simplify
-			"gvif"                     // Run full inliner
-			"CTUca[r]LSsTFOtfDnca[r]Iulc" // SSA plus simplify
-		"]"
+		// should have good "compilability" property here.
+
+		"Tpeul"                        // Run functional expression inliner
+		"xa[r]cL"                      // Turn into SSA again and simplify
+		"gvif"                         // Run full inliner
+		"CTUca[r]LSsTFOtfDnca[r]Iulc"  // SSA plus simplify
+
+		"scCTUt"
+		"gvif"                         // Run full inliner
+		"[scCTUt] TOntnfDIul"          // Perform structural simplification
+
 		"jmul[jul] VcTOcul jmul";      // Make source short and pretty
 
 	static char constexpr DefaultYulOptimiserCleanupSteps[] = "fDnTOcmu";