--- a/test/cmdlineTests/mcopy_string_literal_returned_from_function/output
+++ b/test/cmdlineTests/mcopy_string_literal_returned_from_function/output
@@ -22,25 +22,24 @@ object "C_10" {
                     {
                         if callvalue() { revert(0, 0) }
                         if slt(add(calldatasize(), not(3)), 0) { revert(0, 0) }
-                        let _2 := 64
-                        let newFreePtr := add(_1, _2)
+                        let newFreePtr := add(_1, 64)
                         if or(gt(newFreePtr, 0xffffffffffffffff), lt(newFreePtr, _1))
                         {
                             mstore(0, shl(224, 0x4e487b71))
                             mstore(4, 0x41)
                             revert(0, 0x24)
                         }
-                        mstore(_2, newFreePtr)
+                        mstore(64, newFreePtr)
                         mstore(_1, 25)
-                        let _3 := add(_1, 0x20)
-                        mstore(_3, "MCOPY on Cancun vacation.")
-                        let memPos := mload(_2)
+                        let _2 := add(_1, 0x20)
+                        mstore(_2, "MCOPY on Cancun vacation.")
+                        let memPos := mload(64)
                         mstore(memPos, 0x20)
                         let length := mload(_1)
                         mstore(add(memPos, 0x20), length)
-                        mcopy(add(memPos, _2), _3, length)
-                        mstore(add(add(memPos, length), _2), 0)
-                        return(memPos, add(sub(add(memPos, and(add(length, 31), not(31))), memPos), _2))
+                        mcopy(add(memPos, 64), _2, length)
+                        mstore(add(add(memPos, length), 64), 0)
+                        return(memPos, sub(add(add(memPos, and(add(length, 31), not(31))), 64), memPos))
                     }
                 }
                 revert(0, 0)