--- a/test/cmdlineTests/optimizer_array_sload/output
+++ b/test/cmdlineTests/optimizer_array_sload/output
@@ -18,7 +18,6 @@ object "Arraysum_34" {
             {
                 /// @src 0:80:429  "contract Arraysum {..."
                 mstore(64, memoryguard(0x80))
-                let _1 := 4
                 if iszero(lt(calldatasize(), 4))
                 {
                     if eq(0x81d73423, shr(224, calldataload(0)))
@@ -26,26 +25,31 @@ object "Arraysum_34" {
                         if callvalue() { revert(0, 0) }
                         if slt(add(calldatasize(), not(3)), 0) { revert(0, 0) }
                         let var_sum := 0
+                        /// @src 0:175:182  "sum = 0"
+                        var_sum := /** @src 0:80:429  "contract Arraysum {..." */ 0
                         /// @src 0:368:378  "uint i = 0"
                         let var_i := /** @src 0:80:429  "contract Arraysum {..." */ 0
-                        let _2 := sload(0)
+                        let _1 := sload(0)
                         /// @src 0:364:423  "for(uint i = 0; i < values.length; i++)..."
                         for { }
-                        /** @src 0:380:397  "i < values.length" */ lt(var_i, _2)
-                        /// @src 0:368:378  "uint i = 0"
+                        /** @src 0:368:378  "uint i = 0" */ 1
                         {
                             /// @src 0:399:402  "i++"
                             var_i := /** @src 0:80:429  "contract Arraysum {..." */ add(/** @src 0:399:402  "i++" */ var_i, /** @src 0:80:429  "contract Arraysum {..." */ 1)
                         }
                         /// @src 0:399:402  "i++"
                         {
+                            /// @src 0:380:397  "i < values.length"
+                            let _2 := iszero(lt(var_i, _1))
+                            if _2 { break }
                             /// @src 0:80:429  "contract Arraysum {..."
+                            _2 := 0
                             mstore(0, 0)
                             let sum := add(var_sum, sload(add(18569430475105882587588266137607568536673111973893317399460219858819262702947, var_i)))
                             if gt(var_sum, sum)
                             {
                                 mstore(0, shl(224, 0x4e487b71))
-                                mstore(_1, 0x11)
+                                mstore(4, 0x11)
                                 revert(0, 0x24)
                             }
                             /// @src 0:407:423  "sum += values[i]"