… test/cmdlineTests/standard_debug_info_in_evm_asm_via_ir_location/output.json
… test/cmdlineTests/standard_debug_info_in_yul_location/output.json
… test/cmdlineTests/standard_optimizer_generatedSources/output.json
… test/libsolidity/semanticTests/abiEncoderV1/abi_decode_v2_storage.sol
… test/libsolidity/semanticTests/abiEncoderV1/abi_encode_calldata_slice.sol
… test/libsolidity/semanticTests/abiEncoderV1/struct/struct_storage_ptr.sol
… test/libsolidity/semanticTests/abiEncoderV2/abi_encode_calldata_slice.sol
… test/libsolidity/semanticTests/abiEncoderV2/abi_encode_v2.sol
… test/libsolidity/semanticTests/abiEncoderV2/abi_encode_v2_in_function_inherited_in_v1_contract.sol
… test/libsolidity/semanticTests/abiEncoderV2/calldata_array.sol
… test/libsolidity/semanticTests/abiEncoderV2/calldata_overlapped_dynamic_arrays.sol
… test/libsolidity/semanticTests/abiEncoderV2/storage_array_encoding.sol
… test/libsolidity/semanticTests/abiencodedecode/abi_decode_simple_storage.sol
… test/libsolidity/semanticTests/array/array_memory_index_access.sol
… test/libsolidity/semanticTests/array/array_storage_index_access.sol
… test/libsolidity/semanticTests/array/array_storage_index_boundary_test.sol
… test/libsolidity/semanticTests/array/array_storage_index_zeroed_test.sol
… test/libsolidity/semanticTests/array/array_storage_length_access.sol
… test/libsolidity/semanticTests/array/array_storage_push_empty.sol
… test/libsolidity/semanticTests/array/array_storage_push_empty_length_address.sol
… test/libsolidity/semanticTests/array/array_storage_push_pop.sol
… test/libsolidity/semanticTests/array/arrays_complex_from_and_to_storage.sol
… test/libsolidity/semanticTests/array/byte_array_transitional_2.sol
… test/libsolidity/semanticTests/array/bytes_length_member.sol
… test/libsolidity/semanticTests/array/constant_var_as_array_length.sol
… test/libsolidity/semanticTests/array/copying/array_copy_calldata_storage.sol
… test/libsolidity/semanticTests/array/copying/array_copy_cleanup_uint40.sol
… test/libsolidity/semanticTests/array/copying/array_copy_clear_storage.sol
… test/libsolidity/semanticTests/array/copying/array_copy_different_packing.sol
… test/libsolidity/semanticTests/array/copying/array_copy_including_array.sol