Related to https://github.com/ethereum/solidity/issues/14406

**Do not merge**. This is just a test PR to evaluate how the `the-good-parts` sequence would affect costs in CI.

The sequence is a variant of the [previously tested `single-pass` sequence](https://github.com/ethereum/solidity/pull/14887), but keeping only the parts that seemed to give positive results and with an extra repeated part to keep up with cases where multiple repetitions were still giving the `default` sequence an edge in bytecode size.