@cameel 2024-03-06
I ran my seqbench on `erc20.sol` and looks like the 4% gas increase is coming mostly from a ~65 byte difference in bytecode size:

|                  | runtime gas before |     runtime gas after | bytecode size before |  bytecode size after |
|:-----------------|-------------------:|----------------------:|---------------------:|---------------------:|
| `default`        |            323,363 |       309,700 (-4.2%) |                2,288 |       1,442 (-37.0%) |
| `the-good-parts` |            323,363 |       309,956 (-4.1%) |                2,288 |       1,505 (-34.2%) |

_Note that it's less than 4% here, because `gas_diff_stats.py` compares with the values achieved by the `default` sequence (since these are the ones we have in the repo) and they're smaller than the unoptimized values that seqbench takes as a baseline._

Here's how it looks like over the whole run:

![bytecode-size](https://github.com/ethereum/solidity/assets/137030/b1cca0fd-ea0c-4662-ab5b-301dc7ea5a51)
![runtime-gas](https://github.com/ethereum/solidity/assets/137030/32a0b3c4-ae7a-4e30-b17c-3d98c5caf2dc)